delete user handler

+++ src/routes/user/signup.rs
@@ -1,9 +1,9 @@
 use crate::{
-    models::user::{AccountType, User},
+    models::user::{self, AccountType, User},
     routes::typesense::schema_control::{create_document_schema, update_api_key},
-    utilities::{errors::UserError},
-    utilities::token_wrapper::{TypesenseSecret, NoReplyEmailId, NoReplySecret, NoReplyServer},
-    utilities::email::{send_signup_confirmation},
+    utilities::email::send_signup_confirmation,
+    utilities::errors::UserError,
+    utilities::token_wrapper::{NoReplyEmailId, NoReplySecret, NoReplyServer, TypesenseSecret},
 };
 use axum::{
     extract::{Json, Path, State},
@@ -11,20 +11,20 @@ use axum::{
 };
 use bcrypt::{hash, DEFAULT_COST};
 use chrono::{DateTime, Utc};
-use http::StatusCode;
+use http::{HeaderMap, StatusCode};
 use serde_json::json;
 use sqlx::{Pool, Postgres, Row};
 use validator::Validate;
 
-use super::{generate_token, SignUpForm, TokenResponse, UpdateUser};
+use super::{generate_token, validate_token, SignUpForm, TokenResponse, UpdateUser};
 
 pub async fn create_new_user(
     State(pool): State<Pool<Postgres>>,
     State(jwt_secret): State<String>,
     State(typesense_secret): State<TypesenseSecret>,
     State(no_reply_email_id): State<NoReplyEmailId>,
-    State(no_reply_secret):  State<NoReplySecret>,
-    State(no_reply_server):  State<NoReplyServer>,
+    State(no_reply_secret): State<NoReplySecret>,
+    State(no_reply_server): State<NoReplyServer>,
     Json(form): Json<SignUpForm>,
 ) -> impl IntoResponse {
     form.validate()?;
@@ -97,7 +97,13 @@ pub async fn create_new_user(
 
     //TODO: CONFIRMATION LINK FOR SIGNUP NEEDS TO BE DONE
     let confirmation_link: String = String::from("TEST");
-    send_signup_confirmation(&email, &confirmation_link, &no_reply_email_id.0, &no_reply_secret.0, &no_reply_server.0);
+    send_signup_confirmation(
+        &email,
+        &confirmation_link,
+        &no_reply_email_id.0,
+        &no_reply_secret.0,
+        &no_reply_server.0,
+    );
 
     Ok((StatusCode::OK, serde_json::to_string(&res).unwrap()))
 }
@@ -166,10 +172,23 @@ pub async fn get_users(State(pool): State<Pool<Postgres>>) -> impl IntoResponse 
 }
 
 pub async fn delete_user(
-    Path(id): Path<i64>,
     State(pool): State<Pool<Postgres>>,
-) -> Result<(), UserError> {
-    let q = r#"DELETE FROM userdb.users WHERE id = $1"#;
-    sqlx::query(q).bind(id).execute(&pool).await?;
-    Ok(())
+    State(jwt_secret): State<String>,
+    headers: HeaderMap,
+) -> impl IntoResponse {
+    let auth_header = headers.get("Authorization").unwrap();
+    let jwt = auth_header.to_str().unwrap().replace("Bearer ", "");
+    if let Ok(claims) = validate_token(&jwt, &jwt_secret) {
+        let user_id = claims.sub;
+        let q = r#"DELETE FROM userdb.users WHERE id = $1"#;
+        let result = sqlx::query_as::<_, User>(q)
+            .bind(user_id)
+            .fetch_one(&pool)
+            .await;
+        match result {
+            Ok(_) => return Ok((StatusCode::OK, Json(json!({"message": "user deleted"})))),
+            Err(_) => return Err(UserError::Unauthorized("User not found".to_string())),
+        };
+    }
+    Err(UserError::Unauthorized("Wrong token".to_string()))
 }
